+++ CHANGELOG.md
@@ -21,6 +21,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Battery drain while your server is unreachable, such as when you're away from your home network
+
 ### Other Notes & Contributions
 
 ## [1.2.0]